+++ src/main/java/com/samaxes/maven/minify/plugin/MinifyMojo.java
@@ -24,130 +24,117 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.Parameter;
 
 /**
  * Goal for combining and minifying CSS and JavaScript files.
- *
- * @goal minify
- * @phase process-resources
  */
+@Mojo(name = "minify", defaultPhase = LifecyclePhase.PROCESS_RESOURCES)
 public class MinifyMojo extends AbstractMojo {
 
     /**
      * Webapp source directory.
-     *
-     * @parameter expression="${webappSourceDir}" default-value="${basedir}/src/main/webapp"
      */
+	@Parameter(property = "webappSourceDir", defaultValue = "${basedir}/src/main/webapp")
     private String webappSourceDir;
 
     /**
      * Webapp target directory.
-     *
-     * @parameter expression="${webappTargetDir}" default-value="${project.build.directory}/${project.build.finalName}"
      */
+	@Parameter(property = "webappTargetDir", defaultValue = "${project.build.directory}/${project.build.finalName}")
     private String webappTargetDir;
 
     /**
      * CSS source directory.
-     *
-     * @parameter expression="${cssSourceDir}" default-value="css"
      */
+	@Parameter(property = "cssSourceDir", defaultValue = "css")
     private String cssSourceDir;
 
     /**
      * JavaScript source directory.
-     *
-     * @parameter expression="${jsSourceDir}" default-value="js"
      */
+	@Parameter(property = "jsSourceDir", defaultValue = "js")
     private String jsSourceDir;
 
     /**
      * CSS source filenames list.
-     *
-     * @parameter expression="${cssSourceFiles}" alias="cssFiles"
      */
+	@Parameter(property = "cssSourceFiles", alias = "cssFiles")
     private ArrayList<String> cssSourceFiles;
 
     /**
      * JavaScript source filenames list.
-     *
-     * @parameter expression="${jsSourceFiles}" alias="jsFiles"
      */
+	@Parameter(property = "jsSourceFiles", alias = "jsFiles")
     private ArrayList<String> jsSourceFiles;
 
     /**
      * CSS files to include. Specified as fileset patterns which are relative to the CSS source directory.
-     *
-     * @parameter expression="${cssSourceIncludes}" alias="cssIncludes"
      * @since 1.2
      */
+	@Parameter(property = "cssSourceIncludes", alias = "cssIncludes")
     private ArrayList<String> cssSourceIncludes;
 
     /**
      * JavaScript files to include. Specified as fileset patterns which are relative to the JavaScript source directory.
-     *
-     * @parameter expression="${jsSourceIncludes}" alias="jsIncludes"
      * @since 1.2
      */
+	@Parameter(property = "jsSourceIncludes", alias = "jsIncludes")
     private ArrayList<String> jsSourceIncludes;
 
     /**
      * CSS files to exclude. Specified as fileset patterns which are relative to the CSS source directory.
-     *
-     * @parameter expression="${cssSourceExcludes}" alias="cssExcludes"
      * @since 1.2
      */
+	@Parameter(property = "cssSourceExcludes", alias = "cssExcludes")
     private ArrayList<String> cssSourceExcludes;
 
     /**
      * JavaScript files to exclude. Specified as fileset patterns which are relative to the JavaScript source directory.
-     *
-     * @parameter expression="${jsSourceExcludes}" alias="jsExcludes"
      * @since 1.2
      */
+	@Parameter(property = "jsSourceExcludes", alias = "jsExcludes")
     private ArrayList<String> jsSourceExcludes;
 
     /**
      * CSS target directory.
-     *
-     * @parameter expression="${cssTargetDir}" default-value="css"
      * @since 1.3.2
      */
+	@Parameter(property = "cssTargetDir", defaultValue = "css")
     private String cssTargetDir;
 
     /**
      * JavaScript target directory.
-     *
-     * @parameter expression="${jsTargetDir}" default-value="js"
      * @since 1.3.2
      */
+	@Parameter(property = "jsTargetDir", defaultValue = "js")
     private String jsTargetDir;
 
     /**
      * CSS output filename.
-     *
-     * @parameter expression="${cssFinalFile}" default-value="style.css"
      */
+	@Parameter(property = "cssFinalFile", defaultValue = "style.css")
     private String cssFinalFile;
 
     /**
      * JavaScript output filename.
-     *
-     * @parameter expression="${jsFinalFile}" default-value="script.js"
      */
+	@Parameter(property = "jsFinalFile", defaultValue = "script.js")
     private String jsFinalFile;
 
     /**
      * The output filename suffix.
-     *
-     * @parameter expression="${suffix}" default-value=".min"
      * @since 1.3.2
      */
+	@Parameter(property = "suffix", defaultValue = ".min")
     private String suffix;
 
     /**
@@ -156,10 +143,9 @@ public class MinifyMojo extends AbstractMojo {
      * <strong>Warning:</strong> when both the options <code>nosuffix</code> and <code>skipMerge</code> are set to
      * <code>true</code>, the plugin execution phase needs to be set to <code>package</code>, otherwise the output files
      * will be overridden by the source files during the packaging.
-     *
-     * @parameter expression="${nosuffix}" default-value="false"
      * @since 1.6.2
      */
+	@Parameter(property = "nosuffix", defaultValue = "false")
     private boolean nosuffix;
 
     /**
@@ -171,10 +157,9 @@ public class MinifyMojo extends AbstractMojo {
      * See the <a href="http://www.iana.org/assignments/character-sets">IANA Charset Registry</a> for a list of valid
      * encoding types.
      * </p>
-     *
-     * @parameter expression="${charset}"
      * @since 1.3.2
      */
+	@Parameter(property = "charset")
     private String charset;
 
     /**
@@ -182,81 +167,71 @@ public class MinifyMojo extends AbstractMojo {
      * option is used in that case to split long lines after a specific column. It can also be used to make the code
      * more readable, easier to debug (especially with the MS Script Debugger). Specify 0 to get a line break after each
      * semi-colon in JavaScript, and after each rule in CSS. Specify -1 to disallow line breaks.
-     *
-     * @parameter expression="${linebreak}" default-value="-1"
      */
+	@Parameter(property = "linebreak", defaultValue = "-1")
     private int linebreak;
 
     /**
      * JAVASCRIPT ONLY OPTION!<br/>
      * Minify only. Do not obfuscate local symbols.
-     *
-     * @parameter expression="${munge}" default-value="false"
      */
+	@Parameter(property = "munge", defaultValue = "false")
     private boolean nomunge;
 
     /**
      * JAVASCRIPT ONLY OPTION!<br/>
      * Display informational messages and warnings.
-     *
-     * @parameter expression="${verbose}" default-value="false"
      */
+	@Parameter(property = "verbose", defaultValue = "false")
     private boolean verbose;
 
     /**
      * JAVASCRIPT ONLY OPTION!<br/>
      * Preserve unnecessary semicolons (such as right before a '}'). This option is useful when compressed code has to
      * be run through JSLint (which is the case of YUI for example).
-     *
-     * @parameter expression="${preserveAllSemiColons}" default-value="false"
      */
+	@Parameter(property = "preserveAllSemiColons", defaultValue = "false")
     private boolean preserveAllSemiColons;
 
     /**
      * JAVASCRIPT ONLY OPTION!<br/>
      * Disable all the built-in micro optimizations.
-     *
-     * @parameter expression="${disableOptimizations}" default-value="false"
      */
+	@Parameter(property = "disableOptimizations", defaultValue = "false")
     private boolean disableOptimizations;
 
     /**
      * Size of the buffer used to read source files.
-     *
-     * @parameter expression="${bufferSize}" default-value="4096"
      */
+	@Parameter(property = "bufferSize", defaultValue = "4096")
     private int bufferSize;
 
     /**
      * Maximum execution time in seconds.
-     *
-     * @parameter expression="${timeout}" default-value="30"
      * @since 1.5
      */
+	@Parameter(property = "timeout", defaultValue = "30")
     private long timeout;
 
     /**
      * Show source file paths in log output.
-     *
-     * @parameter expression="${debug}" default-value="false"
      * @since 1.5.2
      */
+	@Parameter(property = "debug", defaultValue = "false")
     private boolean debug;
 
     /**
      * Skip the merge step. Minification will be applied to each source file individually.
-     *
-     * @parameter expression="${skipMerge}" default-value="false"
      * @since 1.5.2
      */
+	@Parameter(property = "skipMerge", defaultValue = "false")
     private boolean skipMerge;
 
     /**
      * Skip the minify step. Useful when merging files that are already minified.
-     *
-     * @parameter expression="${skipMinify}" default-value="false"
      * @since 1.5.2
      */
+	@Parameter(property = "skipMinify", defaultValue = "false")
     private boolean skipMinify;
 
     /**
@@ -271,10 +246,9 @@ public class MinifyMojo extends AbstractMojo {
      * <li><code>yui</code> - <a href="http://developer.yahoo.com/yui/compressor/">YUI Compressor</a></li>
      * <li><code>closure</code> - <a href="https://developers.google.com/closure/compiler/">Google Closure Compiler</a></li>
      * </ul>
-     *
-     * @parameter expression="${jsEngine}" default-value="yui"
      * @since 1.6
      */
+	@Parameter(property = "jsEngine", defaultValue = "yui")
     private String jsEngine;
 
     /**
@@ -297,11 +271,22 @@ public class MinifyMojo extends AbstractMojo {
 
         ExecutorService executor = Executors.newFixedThreadPool(2);
         try {
-            executor.invokeAll(processFilesTasks);
+            final Iterable<Future<Object>> futures = executor.invokeAll(processFilesTasks);
+            checkFuturesResult(futures);
             executor.shutdown();
             executor.awaitTermination(timeout, TimeUnit.SECONDS);
         } catch (InterruptedException e) {
             getLog().error(e.getMessage(), e);
+        }
+    }
+
+    private void checkFuturesResult(final Iterable<Future<Object>> futures) throws MojoExecutionException {
+        for (final Future<Object> future : futures) {
+            try {
+                future.get();
+            } catch (final Exception e) {
+                throw new MojoExecutionException(e.getMessage(), e);
+            }
         }
     }
 }