+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Minify Maven Plugin
 
+## 1.6.3
+* Build should fail if compiler can't parse/compile source files
+* Extends logging
+
 ## 1.6.2
 
 * Cherry picked from commit 31fe5c91bf2d24c29251595206c3c4ebada1c712: Added `nosuffix` option to avoid the suffix `.min` on the minified output file name (#16).